Clean-up of code not necessary with Rails 5.2+
Now that we have all the information we need in the AST, we don't need to carry bindings around.
Since we're still supporting Rails 4.2-5.0-5.1, we can't just delete the code but I added some comments to remind ourselves to delete the code.
I also found that some of the visitor methods won't be necessary anymore.